Finish inventar

+++ inventar/src/inventar/Sellable.java
@@ -1,8 +1,9 @@
 package inventar;
 
 public interface Sellable {
+	//Alle Interfaces angelegt
 	public abstract String getLabel() ;
-	public abstract double getPriceCent();
+	public abstract int getPriceCent();
 	public abstract boolean isAvalibale();
 	public abstract void sellOne();
